rename ut_diagn.py to diagnostics.py
