+++ utide/_solve.py
@@ -8,7 +8,7 @@ import numpy as np
 import scipy       # This will go away; see FIXME below
 
 from .harmonics import ut_E
-from .ut_diagn import ut_diagn
+from .diagnostics import ut_diagn
 from .ut_cs2cep import ut_cs2cep
 from .constituent_selection import ut_cnstitsel
 from .confidence import ut_confidence
